Move request-related functions to a separate module
