+++ src/apify/scrapy/__init__.py
@@ -1,2 +1,3 @@
+from .requests import to_apify_request, to_scrapy_request
 from .scheduler import ApifyScheduler
-from .utils import get_basic_auth_header, get_running_event_loop_id, open_queue_with_custom_client, to_apify_request, to_scrapy_request
+from .utils import get_basic_auth_header, get_running_event_loop_id, open_queue_with_custom_client
