+++ src/apify/scrapy/scheduler.py
@@ -14,7 +14,8 @@ except ImportError as exc:
 
 from apify._crypto import crypto_random_object_id
 from apify.actor import Actor
-from apify.scrapy.utils import nested_event_loop, open_queue_with_custom_client, to_apify_request, to_scrapy_request
+from apify.scrapy.requests import to_apify_request, to_scrapy_request
+from apify.scrapy.utils import nested_event_loop, open_queue_with_custom_client
 from apify.storages import RequestQueue
